organised the Openings Search related components into one folder
