+++ frontend/src/components/SilvicultureSearch/Openings/OpeningsSearchTab/styles.scss
