+++ frontend/src/components/SilvicultureSearch/Openings/SearchFilterBar/SearchFilterBar.scss
