+++ frontend/src/components/SilvicultureSearch/Openings/SearchScreenDataTable/styles.scss
